Make testsuite working in MVP mode
Simple testsuite subsystem which allows to develop tests quickly. No need to repeat boilerplace code again and again.
One just has to write the stuff which matters and use TEST_SUCCESS and TEST_FAIL functions to signalize success or
failure of the test.

Fix some bugs in CMake files which prevented tests to be build elsewhere than in root of the tree