Add support for staged tests
Staged tests allow to check if test passes certain checkpoints in given order. You can
use TEST_STEP function to assert if steps are passed in given order. Function will yield
test failure if the order is not correct and no action if order is correct. The last step
shall be marked with TEST_SUCCESS().