Testsuite now uses project-provided OpenOCD configuration
Testsuite drops the use of its own openocd.cfg. Instead of it it expects that
the project will provide its own OpenOCD configuration that can be used by
the debugger. This way the tests shall be fully portable to whatever configuration
that can provide testing-platform library and openocd.cfg file.