Add step to check if application is sound
One nasty bug to debug is if you make a typo or simply misname application
in the OS_APPLICATION(...) call. Application instance will be defined but
later it will be dropped. Step that checks for correct application instance
name added. It simply checks if correct symbol in correct section exists.