Add HIL test doing basic kill() call test
This test tests if the thread created with higher priority than the
calling thread stops, if it sends SIGSTOP itself. The expected result
is that the thread is stopped and control is given to the calling thread.

If thread continues running after it sent SIGSTOP itself, test is
considered as failed.