Fix API mismatch userspace vs. kernel
Userspace definition of the thread_join() was different than the kernel
implementation of the syscall. The latter returned the exit status code
in function return value, while the userspace implementation expected it
to be written into memory pointed to by the 2nd argument.

The way the kernel does it was selected to be a preferred way as the
pointer approach requires us to actually go and verify that the piece of
memory userspace points at is owned by the process calling the syscall.