Deal with circular dependencies
ARM support code calls some code which now resides in the portable part
of the kernel. This creates circular dependency which linker somehow can't
resolve. The "proper" solution is to use --start-group and --end-group with
all the libraries which belong to the kernel so they behave as if kernel
was just one library. Older CMake versions don't have any way how to
pass this idiomatically correct. Custom interface library is used which
injects correct linker arguments to all targets which link to it.