All the kernel in single library
The previous arrangement with `os` and `cmrx_arch` libraries had issues
during linking while both portions were refering to each other. That
confused the linker.

Due to the strong coupling of these two portions - CMRX may be a microkernel
but this microkernel is a monolith internally - it is better to simply
put everything into one single library.