Use persistent OpenOCD instead of launching it with each test
The old approach was to launch OpenOCD in pipe mode with every test.
This is quite slow and some probes are not that stable so sustain
this. Tests then time out randomly. New approach is to launch OOCD
when first test is launched and OpenOCD is not running yet and then
to reuse that instance of OpenOCD.