Turn test stubs into fake architecture
The testing stubs are de-facto a architecture support-level code.
To make things a tad bit more tidier, lets move this piece of code
into location where architecture support is normally located. This
way the kernel code can link against this support code and work
as normally.

Delete the old libopencm3-related code and shim stubs as this HAL
architecture is not supported anymore.