Reactivate the unit testing support
The unit testing support is activated again. This time unit testing
is not an alternative to the standard build, rather implemented as
nested build. This way no special steps have to be taken in order
to build and run unit tests.

Unit tests are split, updated to reflect the new kernel architecture
and new shimless kernel.

More unit tests are written to check the behavior of thread scheduler.