Make more of the kernel to build for unit testing
Previously only small portion of kernel has been built for unit testing
purposes because the rest of it was heavily dependant on ARM Cortex-M
architecture and low level code. As the kernel is split up into platform
independent portion and platform-specific portion, all of the kernel
can be compiled for unit testing.

Parts that are missing can be filled in by the fake testing architecture
support layer where most of the stubs live.