Integrate unit testing with main tree seamlessly
Add switch UNIT_TESTING which enabled nested unit testing build. This
build will build just the kernel with fake testing architecture layer
and will build unit testing test driver.

The unit testing build is built using native compiler as unit tests are
ran on host machine. There are two targets visible in the main build
system which allow interaction with the unit testing build:

- build_tests - this one will build unit tests
- unit_tests - this one will run the nested make test target

Unit tests are also added as one of tests in the master test suite
if they are enabled. They are omitted if UNIT_TESTING=1 was not passed
to the CMake commandline.

Similarly, if unit testing is enabled but there is no HIL testing
configured and/or enabled, unit tests will be the only test executed
by the master test suite.