Make kernel options configurable
Make options configurable from parent project CMakeLists.txt,
so stuff such as SMP and number of cores can be configured by the
project.

Non-default configuration options are transferred to testing project
so testing uses the same presets as the target build of the kernel.

Enable policy CMP0077 so we are able to tell if option has its default
value or has been configured by the user prior including CMRX sources.