Update tests to expect use of transactions
Templates in tests were changed so that multi start - commit atomic
behavior is expected rather than single lock - unlock behavior happens.

The main difference is that transaction locks and unlock on transaction
start and then locks and unlocks again when transaction is commited.

Bulk of the calculation is done unlocked, but this calculation can't
modify the data. It has to perform commit and if this succeeds, it
can modify the data inside the critical section.

Thus templates got one additional step which goes in between start
and commit and has the same data as the step just before transaction
has been started.