API for UP and SMP differentiation. Support for booting other cores.
API is now differentiated for UP configured systems and for SMP systems.
The rationale is that people using UP systems don't need to deal with
details of on which things run. They all run on the only available core.

Alter existing code so that each individual core can be booted on its
own. This way integrator can boot threads in whatever order and using
whatever way its best for him.