os_thread_migrate skeleton & docs
Added skeleton of os_thread_migrate function. It does nothing useful,
as it is not entirely clear if system of this scale really needs it.

Added some documentation, reworked tests and cleared some warnings.

Timing API needs to be adjusted for SMP support and a way of cross-core
notification needs to be developed.