Kernel fixes of scheduler inconsistency
Massive use of isr_notify_object shown some deficiencies in kernel
thread scheduling process:

The switching of a thread is a two-step process. First a suitable thread
is searched for to switch into. If one is found, then CPU scheduling
context structure is filled-in and thread switch is scheduled.

This process is to make sure that no in-flight ISR is being interrupted
by a thread switch and to make sure that thread switch occurs in
controlled environment.

The issue was that kernel made some changes to scheduling state
immediately after thread switch was scheduled assuming that the process
won't be interrupted.

isr_notify_object routinely interrupts syscalls and can cause thread
rescheduling. This caused certain syscalls to be executed as if they
were called by the thread just being swapped-in on the next thread
switch. This was causing catastrophic failures.