Add syscall to retrieve the current CPU frequency
As drivers are running as pretty much standard userspace process, there
is no access to standard HAL facilities that provide information on CPU
frequency.

This change creates unified syscall that provides access to this
information. As of now it is up on the implementor to implement a way
how this information is retrieved from the HAL.