Add getmicrotime syscall
This adds a userspace syscall that provides current scheduler time. The
actual kernel functionality was present for quite some time, just the
syscall that exports the call into userspace was missing.

Adds a guard assert that checks if syscall was actually called from
userspace context. This catches situations when syscalls are being
called from outside of application thread before kernel was even
started.