Notifications not waking anyone are kept pending
Up until now, if notification of object did not wake up anyone, it was
lost. Any subsequent thread waiting for the notification would remain
blocked.

This change adds a feature of pending notification buffer that records
any notifications which did not wake up anyone. Any notification that
did not wake up any thread is recorded into notification buffer.

If thread waits for an object, then notification buffer is searched for
a notification for this particular object. If it is found, then it is
removed from the buffer and wait is skipped entirely.

Another syscall is added that initializes waitable object. This
initialization causes any threads waiting for the object to be unblocked
and any pending notifications for this object be removed from the buffer.

This syscall can be used as a form of notify_all() for any particular object.

Tests updated to reflect and test for the new behavior of notification
engine.

Note: Notification buffer with fixed size implemented inside kernel
memory region as further change (capabilities/ACL) will require waitable
objects to materialize. Once this is done, buffer will be removed and
notification/waiting thread information will be put into waitable object
itself. This will change userland API in an incompatible way.