Integrate clang-tidy
This adds `apply_clang_tidy()` CMake command which allows seamless
application of clang-tidy to any target. This relies on
compile-commands.json being available in the root directory, so this
option is turned on in the main CMakeLists.txt of the CMRX.

Clang-tidy is enabled by default if clang-tidy executable is found. It
can be disabled by setting CMRX_CLANG_TIDY option to false.

Clang-tidy can be invoked by running the `clang-tidy` target. This is
not done automatically during the build for two reasons:
1) clang-tidy is not particularly fast if many checks are enabled
2) currently it reports some weird errors (such as missing system
   headers)