Fix: map file alias creation w/ out-of-tree CMake
Command to create map file alias was lacking path information. When
CMake was executed in build directory as `cmake path/to/source` it was
working without problems. If CMake was invoked as `cmake -B
path/to/binary path/to/source` the build would fail if map file alias
was actually needed.