Fix: Queue implementation concurrency data hazard
When the last item is read from the queue, then the status of the queue
structure will briefly signalize that the queue is full instead of being
empty. This is due to the fact that queue is inherently implemented as a
lock-free structure.

Put empty flag update in front of the cursor update so this data hazard
won't be ever visible from outside.