Refactor: More flexible timers
Current timer infrastructure has only two types of timers which are
hard-coded into timer infrastructure:

* sleep timer
* interval alarm timer

Their semantics is hard-coded and cannot be changed. If any other kind
of timer is needed (e.g. call timeout timer), this can't be implemented.

Update the timer infrastructure in a way that more types of timers can
be added with their specific handling being defined either directly in
the timer subsystem, or later externally via callbacks.

This change provides an enum for timer type, partitioning timers to
non-periodic one-shot timers and to periodically repeating timers.

Sleep timer and interval alarm timer are ported to the new
infrastructure.