Refactor / update tests
Remove local definitions of timer entry type, this is now shared in
timer.h.

Update some asserts that test for some specific internal behavior.

Add some kernel state manipulating functions. These are mostly added for
a scenario that should not happen in the real world but we test for it
anyway.