Wait for object: implement timeouts
Wait for object syscall did contain timeout argument but up until now
the call returned E_NOTAVAIL if timeout was any value different than
zero.

Now timeout value can be specified in the call to wait, which will
configure timer that aborts the wait for notification and returns
E_TIMEOUT instead.

Substantial rework of the notification framework is part of this change
as the actual handling of what happens when any notification arrives is
up on the callback now. This made the callback a mandatory part of the
framework unlike until now, callback might not be provided.