Fix/Change: STATIC_QUEUE declares type rather than variable
The old behavior of STATIC_QUEUE macro was to define a variable. This is
all nice and compact, but scales poorly if you need to access the queue
from more than one compilation unit.

Thus this macro has been changed to macro STATIC_QUEUE_T which declares
a type instead of variable. Thus the variable has to be defined by the
developer.

The advantage of this approach is that you can declare the queue type in
a header and then define / extern variables as needed in modules.

This is a breaking change for all the code which was already relying on
STATIC_QUEUE.

In order to reveal such code early, we don't provide macro for backwards
compatibility.