Transactions: Add commit tracking
Add simple functions to track potential nested transaction committing.
As transaction commit live in the hot path of the kernel, this option
has compile-time CMake option that enables/disables it.

Is is probably only needed in tests and/or while kernel change is being
developed. In all other scenarios it should not bring any reasonable
benefit assuming that the test suite is reasonably complete.

By default transaction commit tracking is turned off.