Idle thread can sleep CPU
Add option that enables CPU sleeping when idle thread is scheduled. This
is performed from within userspace, so the idle thread is still existing
and timing provider has no clue this actually happens.

This option is disabled by default and has to be enabled by the
top-level project.

Do not use this option if your timing provider depends on clock source
that gets gated by WFI (currently nRF53 seems to have problems with
SysTick timer if WFI is used).