Make sleeping CPU portable
Extend porting layer with one additional function that sleeps CPU.

Note: This is kind-of dirty as idle thread is the one who calls this
function. Userspace code should not call porting layer, nor should be
aware of its existence.

We should probably implement tickless kernel with proper power
management as soon as possible.