Extend porting layer API to support arch init
Add two more functions to porting layer that provide a way how to
initialize the architecture and CPU core on CPU bootup. This is used in
FPU support for Cortex-M to enable automatic handling of the FPU
context but can have many other uses.

Architecture support layers are free to define these functions as empty
macros to denote that no such initialization is needed.