ARM Cortex-M: Add support for floating point
Add basic support for floating point handling. This makes the thread
switcher aware of the FPU context presence and this is saved / restored
for whichever thread that is using floating point instructions.

This implementation is using lazy save/restore of the floating point
state. In this arrangement the FP registers are not saved automatically
when entering handler mode. This speeds up exception processing.