SysTick: Sim-friendly init
If systick is initialized using CMSIS-provided function, then it is
automatically activated. In rare cases (such as stepping through the
code in debugger or running in simulated environment) this may cause
unrealistic early triggering of SysTick_Handler, which gets triggered
before the kernel is booted on that core.

This yields a hard fault.

Thus the timing provider now has its own implementation of the SysTick
configuration routine which does not activate the timer immediately.
Timer is only activated once kernel is ready to process timer
interrupts.