Build system: Clean-up and document
Clean-up certain aspects of the build system. Remove the command to
create symbolic link to possible alternate name of the MAP file. Add
configure-time option to adjust the name instead.

This should enable the Windows build to proceed without failing on the
inability to create symbolic link.

Add documentation of the build system. Document the process of
FindCMSIS, CMRX CMake module, CMRX-specific commands and configure-time
options available to the user.