Kernel: Rework the syscall calling convention
Update the way how porting layers are supposed to define the syscall.

The difference is that old way only communicated the syscall ID to the
porting layer. New way provides a way how to communicate arguments as
well. It is up on the porting layer if it uses this piece of information
or not.