Kernel: Change the way how syscalls are declared
The way how syscalls are declared is made even more target-specific. The
original macro that automatically expands to target section is replaced
by function-like macro that can do whatever the target platform needs.

ARM port updated to change this call to what the original code was
doing.

Linux code declares an array and then crafts a function call that is
registered as constructor. This feeds the internal array of syscalls
with newly added syscalls.

Testing code code basically nothing as it is not using any of these
right now.