API: Application/thread definition platform-specific
Change the way how applications and threads are statically allocated. Up
until now this code was hardcoded for all target platforms. Now the
public API calls entities that are expected to be implemented by
platform-specific code.

ARM code does what the old code was doing up until now. It creates
structures that will end up in specific section that gets picked up by
the linker script and its extents are recorded by symbols that are
later used to iterate over it.

Linux code creates an array which is then processed by a generated
function marked as constructor. This causes that the function is called
before the main and can register these items.

Testing code does nothing as it is not using this facility.