Linux port: Add skeleton of the port
Add initial version of the Linux port. This port is based on ucontext
machinery and getcontext/makecontext/switchcontext/swapcontext calls.

Kernel scheduler is based around signal delivery. This makes the code
very portable as there is no assembly nor CPU-specific code in the port
but it is hardly real-time as it depends on the code calling syscalls in
order for signals to be propagated.

So far the code boots its first task and seems to be switching threads.

There are features that are not implemented and which will change the
way how porting layer looks.

Features that don't work yet:
- syscalls
- RPC calls
- memory protection

Kernel should be able to boot using standard logic, create threads as per
specification and then execute at least the first one. Scheduler is
being called by the hardware and might work with switching threads.

WIP