Add stub of Kernel-based mutexes
Add foundational routines for kernel-based mutexes. This is intended to
use kernel for contended path. Happy path - no contention on mutex -
still runs purely in the userspace.

This allows kernel to handle priority inheritance in case of priority
inversion while keeping code fast by not calling the kernel when it is
not needed.

Mutex code does not use the kernel code entirely correctly yet.