feat: Implement futexes on top of notifications
Notification framework is extended with flag-aware versions of calls
which allow modification of call behavior. This allows for
compare-and-wait operation necessary to implement kernel-assisted
futexes.

Another feature added is priority inheritance. This has to be commanded
from userspace (library deals with it automatically) but is
kernel-driven.

Contention-less path remains fully in userspace for futex locking. Futex
unlocking still has to go into kernel space to notify others. With some
added state management this should also be convertible to fully
userspace execution.

Kernel-only mutex syscall entries were removed as with futexes these
make no sense to exist anymore.

Currently this feature is broken for Cortex M0+ which doesn't support
conditional load/store. The planned resolution of this issue is to treat
futexes on M0+ as fully kernel-only as if __futex_lock_fast() and
__futex_unlock_fast() always failed. This will need some additional
logic on kernel side which is less bad than adding new M0+ exclusive
syscalls effectively duplicating the amount of code.