feat: Inherit thread priority if asked for and waiting
If thread is going to wait for the value of the object to become as
expected then it has possibility to command the kernel to inherit its
scheduling priority to a specified thread.

If this is requested then kernel will check that the target thread is a
valid one and will reconfigure its priority.