clang-tidy: Do not pass mgeneral-regs-only to Clang
Clang-tidy emits an error that -mgeneral-regs-only is not a valid
option for arm-none-eabi target. There is no way to suppress this flag
for clang-tidy.

Thus the approach is to configure the action to configure the build for
Clang and then selectively disable adding this flag if Clang is the
compiler.

Note that Clang is *not* supported as compiler for normal builds yet
even if it was we would have to implement this condition as not even
standard Clang accepts this switch for arm-none-eabi target.