Add coverage reporting for tests
If production components (now the portable part of the OS kernel and the
queue library) are build for testing purposes then these components are
build with code coverage instrumentation turned on. All targets are
linked with libgcov and when these tests are ran then coverage report is
generated.

HTML version of the report can be generated using `coverage` target that
is added to the build.

NB: Current test coverage is roughly 74% of kernel and 98% of queue
server.