Fix: Thread entrypoint argument is invalid for booted thread
First executed thread is entered using software process that emulates
ISR return. This is done in order to avoid having different setup for
first thread and other threads.

The assembly routine which performed this "ISR return" was broken and
popped register values from stack in wrong order. This caused that value
which should end up in R0 (1st argument to the entry point) ended up
somewhere else.

This caused the thread entry point function to work with random value
instead of specified argument if and only if it was the thread that was
the first to start. It did not happen if thread was first started via
thread switcher.