feat: Add support for platform-specific code
Not all chips are made equal and some need special code to handle their
uniqueness. This is a beginning of effort  to allow chip-specific code
in platform support code and chip-specific CMake modules.

Primary purpose of this code is to allow platform specific code for
system, core and thread initialization in a similar way how architecture
can do custom initialization at these points.

Secondary purpose might be to help processing HAL integration. While HAL
integration is not tremendously hard, it is a repetitive work and could
be modularized and automated.