feat: RP2xxx platform support STUB
This is a stub of platform support code for RP2xxx chips. Its purpose is
to create additional MPU region with read-only + execute privileges
which provides access to jump tables used by C runtime and some runtime
functions which are put into RAM rather than FLASH by the linker script
to speed-up certain operations.

This platform code is not fully done yet, it depends on the fact that
linker script contains two additional symbols
__rp2xxx_platform_custom_start and __rp2xxx_platform_custom_end. These
symbols wrap deployments of .text*, .rodata* and .data* inside data
section. Symbols must be aligned to base at least of the length of
region in between them.

Linker script manipulator which would do this is not written yet.