Fix: Rework Linux thread switching mechanism
The original concept of thread switching was based on context switching.
The prototype implementation was working yet it was relying on undefined
behavior of switching context inside signal handler. It turned out it
doesn't work reliably.

Thus the thread switching mechanism was reworked to use Linux threads
and enforce synchronization via sending signals.

This is actually a less hacky and more portable approach + it allows
debugging of individual CMRX threads by host GDB as threads are mapped
1:1.

Port was extensively documented.