Fix: Make syscalls run in "kernel context"
In the old version of the code, syscalls were executed in the same
context as the user code. While there is no memory isolation implemented
in this port yet, this had no major effect except of the possibility
that multiple threads could be in the kernel "context" at the same time
and thread could be scheduled out of the CPU while being in kernel.

This is not supported by the CMRX kernel as the only scenario where
multiple threads are in the kernel context at the same time is if the
machine has multiple cores.

Thus the system call entry mechanism was rewritten in a way that system
calls are dispatched to the "kernel thread". This way, only one thread
can ever only be inside the kernel context at the same time.