Fix: RPC types mismatches
RPC types such as VTable_t and RPC_Method_t are pointer-to-something
kind of types. These were used in the code as VTable_t *, which
constructed double-pointers. As these pointers are not used directly in
ARM port, it wasn't visible and most probably did not cause any
problems, but it is not correct.

So all uses of these types are updated to remove another level of
indirection.