Feature: Implement RPC call syscall on Linux port
This provides bulk of the RPC call system call for the Linux port.

If system call is cleared by the kernel, system call structure is filled
with the command to dispatch the call immediately after system call
returns.

There are still gaps in other parts of infrastructure but change of
VTABLE macro makes this mechanism mostly working.

There is an opportunity to revisit how RPC calls/return syscalls are
implemented. Currently, they are implemented entirely in the porting
layer yet it turns out that some actions may be generalized and provided
as porting layer API while the general logic of the call is kept in the
portable portion of the code.