Feature: Finish RPC call implementation
Add missing implementation of the RPC return system call and finish the
implementation of RPC machinery so that after every dispatched RPC call,
there is RPC return system call automatically performed. This will
overwrite the system call structure in the thread but it doesn't matter
as it was already processed.

The only downside of this approach is that it is done in the userspace,
so application can technically circumvent it.