Feature: Allow renaming of CMRX API symbols
In rare cases where CMRX meets some library that uses same symbol names
as CMRX standard library (e.g. the POSIX Linux port) it is necessary to
resolve this clash.

The way to do it is to define macro CMRX_VERBOSE_API_NAMES which will
cause the symbols generated by the code to have CMRX-specific prefix.
There are macros which override these salted names, so the code looks
the same and works the same as long as correct headers are included.

This mechanism works as long as one piece of code is not trying to call
both CMRX API and the other offending API in the same module.