Fix: Make delivering signals more robust
Signals related to CMRX port have to be delivered into thread that
executes the currently running CMRX thread. Add function that checks for
this and use it all around the code.

Due to the semantics of POSIX signals, we have to account for a case
that "PendSV" is called twice and we are actually trying to switch a
thread for itself. If that happens, then simply ignore the command.

Trigger thread switch even from IRQ, if asked for.