Feature: Peripheral emulator
The peripheral emulator is a feature that allows relatively simple
emulation of peripherals that hooks reads and writes to/from
peripherals. It relies on usage of _R() and _W() macros, so it is really
only usable for environments where emulated peripherals are used.

Peripheral emulator developer can define callbacks for when register is
read or written and these callbacks are then called automatically.

Auxiliary macros will create the emulator object automatically,
initializing callbacks to correct values.